sverak_params: Tune values and refine in original type
Previously it would only refine in Float64 and then convert to the
requested type. Now it refines in the requested type. In particular
this means that it runs a few Newton iterations for Arb.

Also tuned the values for j = 1, d = 3 so that they give better
results for latest version of refine_approximation.